Reorganize lib folder
