+++ lib/domain/SyntaxAggregator.js
