+++ lib/domain/SyntaxModel.js
