+++ lib/domain/SyntaxModes.js
