+++ lib/editor/EditorInserter.js
