+++ lib/editor/KeybindFinder.js
