+++ lib/main.js
@@ -1,18 +1,18 @@
 'use babel';
 
-import SyntaxSearchView from './SyntaxSearchView';
-import SyntaxFilterView from './SyntaxFilterView';
-import SyntaxResultView from './SyntaxResultView';
+import SyntaxSearchView from './view/SyntaxSearchView';
+import SyntaxFilterView from './view/SyntaxFilterView';
+import SyntaxResultView from './view/SyntaxResultView';
 import {
     SyntaxResultTab,
     SyntaxResultSyntaxTab,
     SyntaxResultNotesTab,
     SyntaxResultExampleTab,
     SyntaxResultDocumentationTab,
-} from './SyntaxResultTabs';
+} from './view/tab/SyntaxResultTabs';
 
-import SyntaxSearch from './SyntaxSearch';
-import SyntaxAggregator from './SyntaxAggregator';
+import SyntaxSearch from './domain/SyntaxSearch';
+import SyntaxAggregator from './domain/SyntaxAggregator';
 
 import { CompositeDisposable } from 'atom';
