+++ lib/util/NewLine.js
@@ -1,6 +1,6 @@
 'use babel';
 
-export default class SyntaxUtil {
+export default class NewLine {
     static newLineizeEditorText(text) {
         if (text[text.length - 1] != '\n') {
             return text + '\n';
