+++ lib/util/PercentEncoder.js
