+++ lib/view/SyntaxFilterView.js
@@ -2,8 +2,8 @@
 
 AtomSpacePenViews = require('atom-space-pen-views');
 SelectListView = AtomSpacePenViews.SelectListView;
-import { LanguageFilterMode } from './SyntaxModes';
-import SyntaxCSS from './SyntaxCSS';
+import { LanguageFilterMode } from '../domain/SyntaxModes';
+import SyntaxCSS from './style/SyntaxCSS';
 import { Emitter } from 'atom';
 
 export default class SyntaxFilterView extends SelectListView {
