+++ lib/view/SyntaxResultView.js
@@ -3,7 +3,7 @@
 import { TextEditor } from 'atom';
 import { $, View, TextEditorView } from 'atom-space-pen-views';
 import { CompositeDisposable, Emitter } from 'atom';
-import KeybindFinder from './KeybindFinder';
+import KeybindFinder from '../editor/KeybindFinder';
 
 export default class SyntaxResultView extends View {
     static content() {
