+++ lib/view/SyntaxSearchView.js
