+++ lib/view/style/SyntaxCSS.js
