+++ lib/view/tab/SyntaxResultDocumentationTab.js
@@ -1,7 +1,7 @@
 'use babel';
 
 import SyntaxResultTab from './SyntaxResultTab';
-import SyntaxCSS from './SyntaxCSS';
+import SyntaxCSS from '../style/SyntaxCSS';
 import open from 'open';
 import { TextEditor } from 'atom';
 import { $, TextEditorView } from 'atom-space-pen-views';
