+++ lib/view/tab/SyntaxResultExampleTab.js
@@ -1,12 +1,12 @@
 'use babel';
 
 import SyntaxResultTab from './SyntaxResultTab';
-import EditorInserter from './EditorInserter';
-import SyntaxUtil from './SyntaxUtil';
+import EditorInserter from '../../editor/EditorInserter';
+import KeybindFinder from '../../editor/KeybindFinder';
+import NewLine from '../../util/NewLine';
 import { TextEditor } from 'atom';
 import { TextEditorView } from 'atom-space-pen-views';
 import { CompositeDisposable } from 'atom';
-import KeybindFinder from './KeybindFinder';
 
 export default class SyntaxResultExampleTab extends SyntaxResultTab {
     static content() {
@@ -59,7 +59,7 @@ export default class SyntaxResultExampleTab extends SyntaxResultTab {
     }
 
     setText(text) {
-        this.resultEditorView.setText(SyntaxUtil.newLineizeEditorText(text));
+        this.resultEditorView.setText(NewLine.newLineizeEditorText(text));
     }
 
     onShow() {
