+++ lib/view/tab/SyntaxResultSyntaxTab.js
@@ -1,7 +1,7 @@
 'use babel';
 
 import SyntaxResultTab from './SyntaxResultTab';
-import SyntaxUtil from './SyntaxUtil';
+import NewLine from '../../util/NewLine';
 import { TextEditor } from 'atom';
 import { TextEditorView } from 'atom-space-pen-views';
 
@@ -28,6 +28,6 @@ export default class SyntaxResultSyntaxTab extends SyntaxResultTab {
     }
 
     setText(text) {
-        this.resultEditorView.setText(SyntaxUtil.newLineizeEditorText(text));
+        this.resultEditorView.setText(NewLine.newLineizeEditorText(text));
     }
 }
