+++ lib/view/tab/SyntaxResultTab.js
