+++ lib/view/tab/SyntaxResultTabs.js
